requirements.txt
flask
flask-cors
gunicorn
pandas==2.3.3
scikit-learn==1.6.1
joblib
numpy
xgboost
psycopg2-binary
